backend/requirements.txt
Django~=3.1.7
djangorestframework~=3.12.2
requests~=2.25.1
drf_yasg~=1.20
whitenoise~=5.2.0
